Add tests and fix __init__.py for asc_model
